requirements 2.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
langchain_groq
langchain_google_genai